Add files via upload
A Tenant Management System is a software application designed to help property managers or landlords manage their rental properties, tenants, and related tasks. Here are some of its typical functionalities:

Tenant Tracking: It keeps track of all tenant information, including contact details, rental agreements, payment history, and more.

Rent Collection: It can automate the process of collecting rent, sending reminders for due payments, and recording payment history.

The tenant management system is a python program that creates a GUI created from tkinter that allows users to easily help landlords manage information regarding their tenants, rooms, and payments. The code was created using some of the principles of OOP such as encapsulation since the code was created with the use of classes, that encapsulates related functions to its respective class. Another principle that was used in the code was Abstraction, by implementing the use of class and functions it makes it easy to understand how the code works even without deep understanding of how each function operates.